[22] Feat : 버튼 기능 구현 및 2진 계산기 전환 구현
<description>
구현한 내용들
- 2진 계산기 전환 toggle 버튼 구현
- clear button 구현
- 숫자 입력 버튼 구현
- 양수/음수 전환 버튼 구현

+++ Calculator/Calculator/Model/DecimalCalculator.swift
@@ -23,8 +23,8 @@ final class DecimalCalculator: Calculator {
         return round(Double(firstNumber / secondNumber) * 1e9) / 1e9
     }
     
-    func formatInput(_ userInput: String) throws -> Float {
-        guard let input = Float(userInput) else {
+    func formatInput(_ userInput: String) throws -> Double {
+        guard let input = Double(userInput) else {
             throw CalculatorError.formatError
         }
         return input
